迁入最新版shadowsocksr-libev

+++ ShadowPath/shadowsocks-libev/src/obfs/obfsutil.c
@@ -2,7 +2,7 @@
 #include <time.h>
 
 #include "obfsutil.h"
-#include "../encrypt.h"
+#include "encrypt.h"
 
 int get_head_size(char *plaindata, int size, int def_size) {
     if (plaindata == NULL || size < 2)
@@ -23,7 +23,7 @@ static uint64_t shift128plus_s[2] = {0x10000000, 0xFFFFFFFF};
 void init_shift128plus(void) {
     if (shift128plus_init_flag == 0) {
         shift128plus_init_flag = 1;
-        uint32_t seed = time(NULL);
+        uint32_t seed = (uint32_t)time(NULL);
         shift128plus_s[0] = seed | 0x100000000L;
         shift128plus_s[1] = ((uint64_t)seed << 32) | 0x1;
     }
